About 30% faster now
