+++ argument.yml
@@ -4,11 +4,11 @@ rgb_feat_model_ckpt: checkpoints/color-model.pth
 MiDaS_model_ckpt: MiDaS/model.pt
 fps: 40
 num_frames: 240
-x_shift_range: [-0.03, -0.02, -0.02]
-y_shift_range: [-0.00, -0.00, -0.02]
-z_shift_range: [-0.07, -0.07, -0.07]
-traj_types: ['straight-line', 'circle', 'circle']
-video_postfix: ['zoom-in', 'swing', 'circle']
+x_shift_range: [-0.03]
+y_shift_range: [-0.00]
+z_shift_range: [-0.07]
+traj_types: ['straight-line']
+video_postfix: ['zoom-in']
 specific: ''
 longer_side_len: 960
 src_folder: image
@@ -16,7 +16,7 @@ depth_folder: depth
 mesh_folder: mesh
 video_folder: video
 load_ply: False
-save_ply: True
+save_ply: False
 inference_video: True
 gpu_ids: 0
 offscreen_rendering: False
